add-movement-and-map

+++ src/Vector2D.cpp
@@ -1,6 +1,6 @@
 #include "Vector2D.h"
 
-Vector2D::Vector2D(float x, float y)
+Vector2D::Vector2D(int x, int y)
 {
     this->x = x;
     this->y = y;
