+++ src/Vector2D.h
@@ -3,9 +3,9 @@
 class Vector2D
 {
     public:
-        double x;
-        double y;
-        Vector2D(float x, float y);
+        int x;
+        int y;
+        Vector2D(int x, int y);
         Vector2D();
 };
 #endif
